Duplicates configuration

+++ app/jobs/merge_duplicates_job.rb
@@ -33,12 +33,18 @@ class MergeDuplicatesJob < ApplicationJob
         return 'Invalid record data, perhaps something has changed or another user has resolved these duplicates.'
       end
 
+      return 'The records are the same record.' if duplicate_record.id == record.id
+
       unless Pundit.policy!(current_user, duplicate_record).destroy?
         return 'You do not have authorization to merge these duplicates.'
       end
 
+      status, message = duplicate_record.can_merge_duplicate?(record)
+      return message unless status
+
       duplicate_record.clean_up_duplicate(record)
       duplicate_record.reload
+
       return nil if duplicate_record.destroy
 
       'Could not remove the unused duplicate record '\
